+++ CHANGES.rst
@@ -22,6 +22,17 @@ sbpy.data
 - Added ``DataClass.__contains__`` to enable `in` operator for ``DataClass``
   objects. [#357]
 
+sbpy.activity
+^^^^^^^^^^^^^
+
+- Added ``VectorialModel.binned_production`` constructor for compatibility with
+  time-dependent production implemented in the original FORTRAN vectorial model
+  code by Festou. [#336]
+
+- Added ``VMResult``, ``VMFragmentSputterPolar``, ``VMParams``,
+  ``VMGridParams``, ``VMFragment``, and ``VMParent`` dataclasses to expose
+  details of ``VectorialModel`` results that may be of interest. [#336]
+
 
 Bug Fixes
 ---------
@@ -35,6 +46,15 @@ sbpy.data
 - Asteroids with A/ designations (e.g., A/2019 G2) are correctly identified by
   ``Names`` as asteroids.  Improved handling of interstellar object (I/)
   designations: they do not parse as cometary or asteroidal. [#334, #340]
+
+
+API Changes
+-----------
+
+sbpy.activity
+^^^^^^^^^^^^^
+
+- ``VectorialModel`` now no longer takes an ``angular_substeps`` parameter. [#336]
 
 
 0.3.0 (2022-04-28)